# nix: use mold linker on Linux