Summary: Pure QOL improvement, not that this was too awful as it stood. But this ultimately makes all builds faster, effectively for free; on my Linux machine, debug link time for jj-cli goes from 3s to 1s.

This was originally submitted as part of https://github.com/martinvonz/jj/pull/1864, and backed out as part of https://github.com/martinvonz/jj/pull/1982, because the default `mold` package did not have the proper "wrapper script" that allowed it to find the right external libraries. Using the `mold-wrapped` package however fixes this.

On top of that, let's go ahead and make `mold` the default linker in the Nix package/flake as well; this should improve the `nix build` speed since a lot of tests need to be linked.

On my machine, the following command, which causes a relink of all tests and the `jj` binary, goes from 1m27s to 45s, nearly halving the time &mdash; which is fairly significant when iterating in the test suite.

```
austin@GANON:~/src/jj$ echo >> lib/src/lib.rs 
austin@GANON:~/src/jj$ time cargo test --workspace --no-run
```