@ilyagr 2024-01-22
I could stamp this if you like, but I'd rather somebody who understands Nix better looked at it.

Non-blocking: I think the main benefit is when testing. It might also make sense to make `nix build` use `nextest` if it isn't already.