@thoughtpolice 2024-01-22
Nix is already using nextest, so I think we're in the clear, there. An approval would be nice though; perhaps @necauqua could also provide one, since it would speed up his work too hopefully.