@thoughtpolice 2024-01-22
Ahhhhhh, I forgot I enabled auto merge, so we'll just do it live. :) I feel pretty safe about it this time, though.