@ilyagr 2024-01-22
Maybe we should suggest a Github feature that pops a warning when you resolve the last thread when auto-merge is on.

I have almost learned to never turn auto-merge on before a PR is approved. Still, I do it sometimes, and I regret it 50% of those times. Makes me chuckle; I guess it feels like a lesson about hubris and human folly.